third_party/WebKit/LayoutTests/editing/selection/click-in-padding-with-multiple-line-boxes-expected.txt

Click above/beneath the word FOO. Notice how the cursor goes in the right place (after the letter F).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Selection is [anchorNode: [object Text](12345) anchorOffset: 0 focusNode: [object Text](12345) focusOffset: 0 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](BAR) anchorOffset: 3 focusNode: [object Text](BAR) focusOffset: 3 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](FOO) anchorOffset: 1 focusNode: [object Text](FOO) focusOffset: 1 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](FOO) anchorOffset: 1 focusNode: [object Text](FOO) focusOffset: 1 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](FOO) anchorOffset: 1 focusNode: [object Text](FOO) focusOffset: 1 isCollapsed: true]
PASS Selection is [anchorNode: [object Text](FOO) anchorOffset: 1 focusNode: [object Text](FOO) focusOffset: 1 isCollapsed: true]
PASS successfullyParsed is true

TEST COMPLETE
